requirements.txt
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
psycopg[binary,pool]==3.2.3
pgvector==0.2.5
openai==1.52.2
pypdf==4.3.1
markdown-it-py==3.0.0
regex==2024.9.11